whatsapp-dashboard-bot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.31.0
selenium==4.15.2
webdriver-manager==4.0.1
pydantic==2.5.0
python-dotenv==1.0.0
Pillow==10.1.0
flask==3.0.0
jinja2==3.1.2
sqlalchemy==2.0.23